fixed navbar, removed css files, added taiwlindcss
