+++ client/package.json
@@ -20,7 +20,7 @@
     "react": "^18.2.0",
     "react-chartjs-2": "^5.3.0",
     "react-dom": "^18.2.0",
-    "react-icons": "^4.10.1",
+    "react-icons": "^4.12.0",
     "react-router-dom": "^7.5.2",
     "recharts": "^2.15.3",
     "tailwind-merge": "^3.2.0"
